Remove outdated version check components and tests, replacing them with a new implementation for version updates. Introduce CheckUpdatePage and UpdatePrompt components for improved user experience in handling updates.
